Squashed commit of the following:
commit 27c9c6b61b5655dfcb629a329abe8d798591578a
Author: DenisKochetov <denkochetov322@gmail.com>
Date:   Fri Jan 14 07:28:25 2022 +0500

    Revert "Works perfectly with real account with real token"

    This reverts commit 37781f4244638f8c33490c568c177287ff0e72a2.

commit 0a0a6910bdb730764f23ff7bae2d5a4b15ec84ce
Author: DenisKochetov <denkochetov322@gmail.com>
Date:   Fri Jan 14 07:27:19 2022 +0500

    Added different intervals for candles

commit 37781f4244638f8c33490c568c177287ff0e72a2
Author: DenisKochetov <denkochetov322@gmail.com>
Date:   Fri Jan 14 01:16:16 2022 +0500

    Works perfectly with real account with real token

commit 014f327ee1ac3428daece2026ee7429c6586494a
Author: DenisKochetov <denkochetov322@gmail.com>
Date:   Fri Jan 14 01:15:10 2022 +0500

    Changed padding for portfolio

commit ccdecb49b691a1437207365405341516e8a9a0ac
Author: DenisKochetov <denkochetov322@gmail.com>
Date:   Fri Jan 14 00:08:32 2022 +0500

    Fixed chart display

    now no need in clicking

commit b76ae97eeb2e6eff66add01441316f652b2fa9f2
Author: DenisKochetov <denkochetov322@gmail.com>
Date:   Thu Jan 13 22:45:14 2022 +0500

    deleted garbage

commit 695b429104bd7b0d99b9735a9049de0f67c9e407
Author: DenisKochetov <denkochetov322@gmail.com>
Date:   Thu Jan 13 22:41:22 2022 +0500

    Added date formatting

commit 94c8c64f6e640c7d3b3e8a4d5571498822059ee2
Author: DenisKochetov <denkochetov322@gmail.com>
Date:   Thu Jan 13 22:08:13 2022 +0500

    Deleted test garbage

commit 03e7c7651e79eb400f279a78d741786f05f38e5e
Author: DenisKochetov <denkochetov322@gmail.com>
Date:   Thu Jan 13 22:02:46 2022 +0500

    Added xlabels

    now needs formatting for dates

commit 69262e785fda2ba53b6f7dcea40cf7468f36cb9c
Author: DenisKochetov <denkochetov322@gmail.com>
Date:   Thu Jan 13 21:24:17 2022 +0500

    Added candleData to candleChart

    works fine, but we need xlabels (time) and description
    also works too slowly

commit 656be4629413e502633130fba22fa873aa9e8265
Merge: c895d2b fa0cfe3
Author: DenisKochetov <denkochetov322@gmail.com>
Date:   Thu Jan 13 17:35:00 2022 +0500

    Merge branch 'main' into feature/candles

commit fa0cfe3a461da723e95cd812b23d047a02885f9d
Author: DenisKochetov <denkochetov322@gmail.com>
Date:   Thu Jan 13 17:34:31 2022 +0500

    Fixed isin for currencies

commit c895d2ba07a2c79e92a575fe3028e01416e9b7c1
Author: DenisKochetov <denkochetov322@gmail.com>
Date:   Thu Jan 13 17:22:36 2022 +0500

    Added stickChart but with errors

commit 0fbfbdfafe1403aba1b5651f0ef6c44b7980e766
Author: Владислав <v.pyslar@nyamee.com>
Date:   Thu Jan 13 14:18:37 2022 +0300

    Default icon stonks

    Убрал проверку на иконку

commit 60c40f96675b7a430f6a541eb4ba1325fb44c5df
Author: Владислав <v.pyslar@nyamee.com>
Date:   Thu Jan 13 14:16:08 2022 +0300

    edit item_list_portfolio

+++ app/src/main/java/com/mipt/android/data/api/responses/CandlesResponse.kt
@@ -8,11 +8,11 @@ data class CandlesResponse(
 ) {
     @Serializable
     data class Candle(
-        val o: Float,
-        val c: Float,
-        val h: Float,
-        val l: Float,
-        val v: Int,
+        val o: Double,
+        val c: Double,
+        val h: Double,
+        val l: Double,
+        val v: Double,
         val time: String,
         val interval: String,
         val figi: String
